Detail Enhance

+++ Keeping/Keeping/KPTimePicker/KPTimePicker.m
@@ -138,7 +138,7 @@
 {
     CGPoint velocity = [sender velocityInView:self];
     CGPoint location = [sender locationInView:self];
-    CGFloat vel = fabsf(velocity.y)+fabsf(velocity.x);
+    CGFloat vel = fabs(velocity.y)+fabs(velocity.x);
     NSInteger minutesPerInterval = 10;
     CGFloat angleInterval = 50;
     //NSLog(@"%f",vel);
